refactor: userReportのフォルダ名をreportに変更 #163
